RENAME: rename entity package
