+++ src/main/java/umc/entity/category/Category.java
@@ -1,12 +1,10 @@
-package umc.Entity.category;
+package umc.entity.category;
 
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Getter
