+++ src/main/java/umc/entity/category/UserCategoryPK.java
@@ -1,4 +1,4 @@
-package umc.Entity.mission;
+package umc.entity.category;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -14,11 +14,11 @@ import java.io.Serializable;
 @AllArgsConstructor
 @Getter
 @EqualsAndHashCode
-public class UserMissionPK implements Serializable {
+public class UserCategoryPK implements Serializable {
 
     @Column(name = "user_id")
     private Long userId;
 
-    @Column(name = "mission_id")
-    private Long missionId;
+    @Column(name = "category_id")
+    private Long categoryId;
 }
