+++ src/main/java/umc/entity/inquiry/Inquiry.java
@@ -1,13 +1,11 @@
-package umc.Entity.inquiry;
+package umc.entity.inquiry;
 
 
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Getter
