+++ src/main/java/umc/entity/store/StoreImage.java
@@ -1,24 +1,28 @@
-package umc.Entity.review;
+package umc.entity.store;
+
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-@Builder
+@Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-@Entity
-@Table(name = "review_image")
-public class ReviewImage {
+@Table(name="store_image")
+public class StoreImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, unique = true)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "review_id", nullable = false)
-    private Review review;
+    @JoinColumn(name = "store_id", nullable = false)
+    private Store store;
 
-    @Column(name = "review_image_url", nullable = false)
+    @Column(name = "store_image_url", nullable = false)
     private String imageUrl;
+
 }
